Add files via upload
# Tic-Tac-Toe Game

This project is a modern implementation of the classic Tic-Tac-Toe game, built with HTML, CSS, and JavaScript. It features a clean UI with animations, and a responsive layout, making it enjoyable to play on various devices. Each game offers an intuitive user experience with clear win conditions and a restart option.

Created as part of my internship at Prodigy Inotech, this project helped strengthen my understanding of front-end development and JavaScript logic, particularly in handling game states and interactivity. The project serves as both a fun game and a solid example of applying web development skills to create engaging applications.